🐛 [BUGFIX] 비밀번호 확인 유효성 검사 버그 수정 #248
